Add a chunk init function
Plus tests for that. It still is a bit awkward: you need to add two to the
chunk data size (to account for the header) and then ql_set_append() with that
size. It might be better to make a ql_set_append_chunk(); routine that takes
the data size instead and builds the chunk header for the user.